refactor: Константы для списков вынесены в файлы
