+++ src/shared/constants/header-social-items.ts
@@ -1,0 +1,24 @@
+import { TSocialItem } from 'components/ui/menu';
+
+export const HEADER_SOCIAL_ITEMS: TSocialItem[] = [
+  {
+    title: 'fb',
+    href: '#'
+  },
+  {
+    title: 'inst',
+    href: '#'
+  },
+  {
+    title: 'ytube',
+    href: '#'
+  },
+  {
+    title: 'tlgrm',
+    href: '#'
+  },
+  {
+    title: 'vk',
+    href: '#'
+  },
+];
